requirements.txt
fastapi[standard]>=0.116.1,<0.117.0
sqlalchemy[asyncio]>=2.0.41,<3.0.0
aiosqlite>=0.19.0,<0.20.0
pydantic-settings>=2.10.1,<3.0.0
alembic>=1.16.4,<2.0.0
pwdlib[argon2]>=0.2.1,<0.3.0
pyjwt>=2.10.1,<3.0.0
tzdata>=2025.2,<2026.0
python-jose[jwt]>=3.4.0,<4.0.0